Tests for CreateProjectFromDatabaseQuickpick
